MS4-T4 - config/packages + Admin/ - Refactor paths to more consistent for admin
